+++ src/main/java/com/d129cm/backendapi/auth/config/MemberSecurityConfig.java
@@ -8,6 +8,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.AuthenticationProvider;
@@ -18,9 +20,9 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
@@ -34,28 +36,26 @@ import java.util.List;
 public class MemberSecurityConfig {
 
     private final JwtProvider jwtProvider;
+    private final PasswordEncoder passwordEncoder;
     private final MemberDetailsService memberService;
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
 
     @Bean
     public AuthenticationProvider memberProvider() {
         DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
-        provider.setPasswordEncoder(passwordEncoder());
+        provider.setPasswordEncoder(passwordEncoder);
         provider.setUserDetailsService(memberService);
         return provider;
     }
+
     @Bean
+    @Primary
     public AuthenticationManager memberAuthenticationManager() {
         return new ProviderManager(memberProvider());
     }
 
 
     @Bean
-    public MemberJwtLoginFilter memberAuthenticationFilter() {
+    public MemberJwtLoginFilter memberJwtLoginFilter() {
         MemberJwtLoginFilter memberJwtLoginFilter = new MemberJwtLoginFilter(jwtProvider);
         memberJwtLoginFilter.setAuthenticationManager(memberAuthenticationManager());
         return memberJwtLoginFilter;
@@ -70,7 +70,6 @@ public class MemberSecurityConfig {
     CorsConfigurationSource corsConfigurationSource() {
         CorsConfiguration configuration = new CorsConfiguration();
         configuration.setAllowedHeaders(List.of("*"));
-//        configuration.setAllowedOrigins(List.of("http://localhost:3000", "https://eroom-challenge.com", "https://www.eroom-challenge.com"));
         configuration.setAllowCredentials(true);
         configuration.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"));
         configuration.setMaxAge(60L);
@@ -81,10 +80,12 @@ public class MemberSecurityConfig {
     }
 
     @Bean(name = "memberSecurityFilterChain")
+    @Order(1)
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.cors(Customizer.withDefaults())
+        http.securityMatcher("/members/**")
+                .cors(Customizer.withDefaults())
                 .csrf(AbstractHttpConfigurer::disable)
-                .sessionManagement(s ->  s.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
+                .sessionManagement(s -> s.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
 
         http.authorizeHttpRequests(common -> common
                 .requestMatchers(PathRequest.toStaticResources().atCommonLocations()).permitAll());
@@ -95,6 +96,9 @@ public class MemberSecurityConfig {
 
         http.formLogin(AbstractHttpConfigurer::disable)
                 .httpBasic(AbstractHttpConfigurer::disable);
+
+        http.addFilterBefore(memberJwtLoginFilter(), UsernamePasswordAuthenticationFilter.class)
+                .addFilterAfter(jwtAuthorizationFilter(), MemberJwtLoginFilter.class);
 
         return http.build();
     }
