+++ src/main/java/com/d129cm/backendapi/auth/domain/MemberDetails.java
@@ -16,7 +16,7 @@ public record MemberDetails (Member member) implements UserDetails {
 
     @Override
     public String getPassword() {
-        return member.getPassword().getValue();
+        return member.getPassword().getPassword();
     }
 
     @Override
